Separate private and public landing requests
Private requests must be claimed before use. Public / Anonymous requests
only depend on the public flag, so anoymous users can create private
landing requests and authenticated users can create public requests.

Tests should cover all of these scenarios.

+++ lib/galaxy/model/migrations/alembic/versions_gxy/a99a5b52ccb8_add_workflow_source_and_workflow_source_.py
@@ -1,4 +1,4 @@
-"""Add workflow_source and workflow_source_type column
+"""Add workflow_source and workflow_source_type and public column
 
 Revision ID: a99a5b52ccb8
 Revises: 7ffd33d5d144
@@ -10,6 +10,7 @@ import sqlalchemy as sa
 
 from galaxy.model.migrations.util import (
     add_column,
+    column_exists,
     drop_column,
     transaction,
 )
@@ -20,17 +21,27 @@ down_revision = "7ffd33d5d144"
 branch_labels = None
 depends_on = None
 
-table_name = "workflow_landing_request"
-column_names = ["workflow_source", "workflow_source_type"]
+workflow_table_name = "workflow_landing_request"
+tool_table_name = "tool_landing_request"
+
+
+def drop_if_exists(table_name: str, column_name: str):
+    if column_exists(table_name, column_name, True):
+        drop_column(table_name, column_name)
 
 
 def upgrade():
     with transaction():
-        for column_name in column_names:
-            add_column(table_name, sa.Column(column_name, sa.String(255), nullable=True))
+        add_column(workflow_table_name, sa.Column("workflow_source", sa.String(255), nullable=True))
+        add_column(workflow_table_name, sa.Column("workflow_source_type", sa.String(255), nullable=True))
+        add_column(workflow_table_name, sa.Column("public", sa.Boolean, nullable=False))
+        add_column(tool_table_name, sa.Column("public", sa.Boolean, nullable=False))
 
 
 def downgrade():
     with transaction():
-        for column_name in column_names:
-            drop_column(table_name, column_name)
+        drop_column(workflow_table_name, "workflow_source")
+        drop_column(workflow_table_name, "workflow_source_type")
+        # For test.galaxyproject.org which was deployed from PR branch that didn't contain public column
+        drop_if_exists(workflow_table_name, "public")
+        drop_if_exists(tool_table_name, "public")
